Moved utils to datasource.service and controlstream.service
